+++ src/main/java/com/fourthread/ozang/module/domain/security/filter/JsonLoginFilter.java
@@ -1,8 +1,8 @@
 package com.fourthread.ozang.module.domain.security.filter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fourthread.ozang.module.domain.security.CustomLoginFailureHandler;
-import com.fourthread.ozang.module.domain.security.CustomLoginSuccessHandler;
+import com.fourthread.ozang.module.domain.security.handler.CustomLoginFailureHandler;
+import com.fourthread.ozang.module.domain.security.handler.CustomLoginSuccessHandler;
 import com.fourthread.ozang.module.domain.security.SecurityMatchers;
 import com.fourthread.ozang.module.domain.user.dto.request.LoginRequest;
 import jakarta.servlet.http.HttpServletRequest;
