+++ src/main/java/com/fourthread/ozang/module/domain/security/handler/CustomAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.fourthread.ozang.module.domain.security;
+package com.fourthread.ozang.module.domain.security.handler;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
