+++ src/main/java/com/fourthread/ozang/module/domain/security/handler/CustomLoginFailureHandler.java
@@ -1,4 +1,4 @@
-package com.fourthread.ozang.module.domain.security;
+package com.fourthread.ozang.module.domain.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fourthread.ozang.module.common.exception.ErrorResponse;
